# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Short Video Maker Import API
  description: |
    Comprehensive video import API for analyzing, downloading, processing, and segmenting videos from various sources.
    
    Features:
    - Video analysis from YouTube, TikTok, Instagram, Facebook, and direct URLs
    - AI-powered content analysis and segmentation
    - Real-time progress updates via WebSocket
    - Multi-language translation support
    - Ollama local AI integration
    
    This API enables importing long-form videos and converting them into short-form content suitable for TikTok, Instagram Reels, and YouTube Shorts.
  version: 1.0.0
  contact:
    name: Short Video Maker Support
    url: https://github.com/your-repo/short-video-maker
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3233/api
    description: Development server
  - url: https://your-domain.com/api
    description: Production server
...
paths:
...
  /translation/languages:
    get:
      tags: [Translation]
      summary: Get supported languages
      description: Retrieve list of supported languages for translation services.
      operationId: getSupportedLanguages
      responses:
        '200':
          description: Supported languages retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationLanguagesResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /translation/translate:
    post:
      tags: [Translation]
      summary: Translate text
      description: |
        Translate text using multiple providers (OpenAI, Google, DeepL) with automatic
        fallback and alternative suggestions.
      operationId: translateText
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranslateRequest'
            examples:
              auto_detect:
                summary: Auto-detect source language
                value:
                  text: "Hello, how are you today?"
                  targetLanguage: "es"
              specific_source:
                summary: Specify source language
                value:
                  text: "Bonjour, comment allez-vous?"
                  sourceLanguage: "fr"
                  targetLanguage: "en"
      responses:
        '200':
          description: Translation completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
...
    TranslationLanguage:
      type: object
      required: [code, name, nativeName]
      properties:
        code:
          type: string
          description: ISO 639-1 language code
          example: "es"
        name:
          type: string
          description: English name of the language
          example: "Spanish"
        nativeName:
          type: string
          description: Native name of the language
          example: "Español"
    TranslationLanguagesResponse:
      type: object
      required: [languages, defaultSource, defaultTarget]
      properties:
        languages:
          type: array
          items:
            $ref: '#/components/schemas/TranslationLanguage'
        defaultSource:
          type: string
          example: "auto"
        defaultTarget:
          type: string
          example: "en"
    TranslateRequest:
      type: object
      required: [text, targetLanguage]
      properties:
        text:
          type: string
          minLength: 1
          description: Text to translate
          example: "Hello, how are you today?"
        targetLanguage:
          type: string
          minLength: 2
          description: Target language code (ISO 639-1)
          example: "es"
        sourceLanguage:
          type: string
          description: Source language code (auto-detect if not provided)
          example: "en"
    TranslationResult:
      type: object
      required: [translatedText, sourceLanguage, targetLanguage, confidence, alternatives]
      properties:
        translatedText:
          type: string
          example: "Hola, ¿cómo estás hoy?"
        sourceLanguage:
          type: string
          example: "en"
        targetLanguage:
          type: string
          example: "es"
        confidence:
          type: number
          minimum: 0
          maximum: 1
          example: 0.98
        alternatives:
          type: array
          items:
            type: string
          example: ["Hola, ¿cómo te encuentras hoy?", "Hola, ¿qué tal estás hoy?"]
        provider:
          type: string
          example: "openai"
    TranslateResponse:
      type: object
      required: [success, translation, translatedAt]
      properties:
        success:
          type: boolean
        translation:
          $ref: '#/components/schemas/TranslationResult'
        translatedAt:
          type: string
          format: date-time
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Error message
        details:
          type: string
          description: Additional error details
        code:
          type: string
          description: Error code for programmatic handling
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            invalid_url:
              summary: Invalid URL format
              value:
                error: "Invalid request data"
                details: "URL must be a valid URL format"
                code: "INVALID_URL"
            validation_error:
              summary: Request validation failed
              value:
                error: "Invalid request data"
                details: "Request body validation failed"
...
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per hour
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in current window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when rate limit resets
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before next request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Rate limit exceeded"
            details: "Too many requests. Please wait before making another request."
            code: "RATE_LIMIT_EXCEEDED"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            analysis_failed:
              summary: Video analysis failed
              value:
                error: "Failed to analyze video"
                details: "Unable to extract video metadata"
                code: "ANALYSIS_FAILED"
            processing_error:
              summary: Processing error
              value:
                error: "Internal server error"
                details: "An unexpected error occurred during processing"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (when required)

# Currently no security is required, but this is defined for future use
security: []
...